refactor: CardDispenser 인터페이스 제거
